% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_form_data.R
\name{get_form_data}
\alias{get_form_data}
\title{Get Form Data}
\usage{
get_form_data(form_url)
}
\arguments{
\item{form_url}{\itemize{
\item the share link for the sheet which the google form drops
data into
}}
}
\value{
a dataframe of lift breakage data
}
\description{
Get Form Data
}
